Added mdoc and wrote up some documentation for `MonadCancel`
