+++ core/shared/src/main/scala/cats/effect/syntax/AllSyntax.scala
@@ -18,7 +18,8 @@ package cats.effect
 package syntax
 
 trait AllSyntax
-    extends kernel.syntax.GenSpawnSyntax
+    extends kernel.syntax.MonadCancelSyntax
+    with kernel.syntax.GenSpawnSyntax
     with kernel.syntax.GenTemporalSyntax
     with kernel.syntax.AsyncSyntax
     with kernel.syntax.SyncEffectSyntax
