+++ project/plugins.sbt
@@ -5,3 +5,4 @@ addSbtPlugin("org.scala-js"       % "sbt-scalajs"              % "1.2.0")
 addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.0.0")
 addSbtPlugin("pl.project13.scala" % "sbt-jmh"                  % "0.3.7")
 addSbtPlugin("org.scalameta"      % "sbt-scalafmt"             % "2.4.0")
+addSbtPlugin("org.scalameta"      % "sbt-mdoc"                 % "2.2.9")
